refactor: 업비트 api응답 시 배열이 아닌 list로 받도록 변경
