+++ indicator-service/src/main/kotlin/com/emgc/indicatorservice/service/upbit/DefaultUpbitCandleApiService.kt
@@ -1,15 +1,16 @@
 package com.emgc.indicatorservice.service.upbit
 
-import com.emgc.indicatorservice.dto.upbit.request.UpbitDayCandleRequest
-import com.emgc.indicatorservice.dto.upbit.request.UpbitMinuteCandleRequest
-import com.emgc.indicatorservice.dto.upbit.request.UpbitMonthCandleRequest
-import com.emgc.indicatorservice.dto.upbit.request.UpbitWeekCandleRequest
-import com.emgc.indicatorservice.dto.upbit.response.UpbitDayCandleResponse
-import com.emgc.indicatorservice.dto.upbit.response.UpbitMinuteCandleResponse
-import com.emgc.indicatorservice.dto.upbit.response.UpbitMonthCandleResponse
-import com.emgc.indicatorservice.dto.upbit.response.UpbitWeekCandleResponse
+import com.emgc.indicatorservice.dto.upbit.candle.request.UpbitDayCandleRequest
+import com.emgc.indicatorservice.dto.upbit.candle.request.UpbitMinuteCandleRequest
+import com.emgc.indicatorservice.dto.upbit.candle.request.UpbitMonthCandleRequest
+import com.emgc.indicatorservice.dto.upbit.candle.request.UpbitWeekCandleRequest
+import com.emgc.indicatorservice.dto.upbit.candle.response.UpbitDayCandleResponse
+import com.emgc.indicatorservice.dto.upbit.candle.response.UpbitMinuteCandleResponse
+import com.emgc.indicatorservice.dto.upbit.candle.response.UpbitMonthCandleResponse
+import com.emgc.indicatorservice.dto.upbit.candle.response.UpbitWeekCandleResponse
 import com.emgc.indicatorservice.util.urlgenerator.UpbitUrlGenerator
 import mu.KLogging
+import org.springframework.core.ParameterizedTypeReference
 import org.springframework.http.MediaType
 import org.springframework.stereotype.Service
 import org.springframework.web.reactive.function.client.WebClient
@@ -20,28 +21,24 @@ class DefaultUpbitCandleApiService(
     private val upbitUrlGenerator: UpbitUrlGenerator
 ): UpbitCandleApiService {
 
-    companion object {
-        private val log = KLogging()
-    }
-
     override fun getMinuteCandle(
         unit: Int,
         request: UpbitMinuteCandleRequest
-    ): Array<UpbitMinuteCandleResponse>? {
+    ): List<UpbitMinuteCandleResponse> {
         val requestUrl = upbitUrlGenerator.getMinuteCandleUrl(unit, request)
-
+        
         return webClient
             .get()
             .uri(requestUrl)
             .accept(MediaType.APPLICATION_JSON)
             .retrieve()
-            .bodyToMono(Array<UpbitMinuteCandleResponse>::class.java)
-            .block()
+            .bodyToMono(object: ParameterizedTypeReference<List<UpbitMinuteCandleResponse>>() {})
+            .block()!!
     }
 
     override fun getDayCandle(
         request: UpbitDayCandleRequest
-    ): Array<UpbitDayCandleResponse>? {
+    ): List<UpbitDayCandleResponse> {
         val requestUrl = upbitUrlGenerator.getDayCandleUrl(request)
 
         return webClient
@@ -49,13 +46,13 @@ class DefaultUpbitCandleApiService(
             .uri(requestUrl)
             .accept(MediaType.APPLICATION_JSON)
             .retrieve()
-            .bodyToMono(Array<UpbitDayCandleResponse>::class.java)
-            .block()
+            .bodyToMono(object: ParameterizedTypeReference<List<UpbitDayCandleResponse>>() {})
+            .block()!!
     }
 
     override fun getWeekCandle(
         request: UpbitWeekCandleRequest
-    ): Array<UpbitWeekCandleResponse>? {
+    ): List<UpbitWeekCandleResponse> {
         val requestUrl = upbitUrlGenerator.getWeekCandleUrl(request)
 
         return webClient
@@ -63,13 +60,13 @@ class DefaultUpbitCandleApiService(
             .uri(requestUrl)
             .accept(MediaType.APPLICATION_JSON)
             .retrieve()
-            .bodyToMono(Array<UpbitWeekCandleResponse>::class.java)
-            .block()
+            .bodyToMono(object: ParameterizedTypeReference<List<UpbitWeekCandleResponse>>() {})
+            .block()!!
     }
 
     override fun getMonthCandle(
         request: UpbitMonthCandleRequest
-    ): Array<UpbitMonthCandleResponse>? {
+    ): List<UpbitMonthCandleResponse> {
         val requestUrl = upbitUrlGenerator.getMonthCandleUrl(request)
 
         return webClient
@@ -77,8 +74,8 @@ class DefaultUpbitCandleApiService(
             .uri(requestUrl)
             .accept(MediaType.APPLICATION_JSON)
             .retrieve()
-            .bodyToMono(Array<UpbitMonthCandleResponse>::class.java)
-            .block()
+            .bodyToMono(object: ParameterizedTypeReference<List<UpbitMonthCandleResponse>>() {})
+            .block()!!
     }
 
 }
